requirements.txt
# License Plate Detection — Python Dependencies
# Install with: pip install -r requirements.txt

# Core
numpy>=1.24.0
opencv-python-headless>=4.8.0
Pillow>=9.0.0
PyYAML>=6.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Deep Learning (YOLOv8)
ultralytics>=8.0.0

# Web Application
streamlit>=1.28.0

# OCR (Optional — for plate text extraction)
easyocr>=1.7.0
